Changed access to row and col to row() and col() in RowCol. Now BinID and RowCol are derived from IdxPair.
git-svn-id: https://svn.opendtect.org/od/trunk@31898 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/MPEEngine/horizon3dextender.cc
@@ -158,7 +158,7 @@ int BaseHorizon3DExtender::nextStep()
 		    const RowCol dst( (RowCol::fromInt64(serc))/step );
 		    const RowCol cursrc( srcbid/step );
 
-		    const int olddist = oldsrc.sqDistTo(dst);
+		    const int olddist = (int)oldsrc.sqDistTo(dst);
 		    if ( cursrc.sqDistTo( dst )<olddist ) 
 		    {
 			addedpossrc_[previndex] = srcbid.toInt64();
